hwkmetrics-507 - make injecting CacheClient easier by moving it and CacheKey to its own jar - add ignoreFiltering option to sendData/Events to allow clients to   perform their own filtering and (java clients only for now)
