+++ hawkular-alerts-engine/pom.xml
@@ -39,6 +39,11 @@
       <artifactId>hawkular-alerts-api</artifactId>
       <version>${project.version}</version>
     </dependency>
+    <dependency>
+      <groupId>org.hawkular.alerts</groupId>
+      <artifactId>hawkular-alerts-filter-api</artifactId>
+      <version>${project.version}</version>
+    </dependency>
 
     <dependency>
       <groupId>org.kie</groupId>
