+++ pom.xml
@@ -125,6 +125,7 @@
       <modules>
         <module>hawkular-alerts-api</module>
         <module>hawkular-alerts-bus-api</module>
+        <module>hawkular-alerts-filter-api</module>
         <module>hawkular-alerts-engine</module>
         <module>hawkular-alerts-bus</module>
         <module>hawkular-alerts-rest</module>
@@ -149,6 +150,7 @@
       </properties>
       <modules>
         <module>hawkular-alerts-api</module>
+        <module>hawkular-alerts-filter-api</module>
         <module>hawkular-alerts-engine</module>
         <module>hawkular-alerts-rest</module>
         <module>hawkular-alerts-actions</module>
